+++ go.mod
@@ -13,6 +13,7 @@ require (
 )
 
 require (
+	bou.ke/monkey v1.0.2
 	github.com/stretchr/testify v1.10.0
 	golang.org/x/sys v0.30.0 // indirect
 )
